Add IO class; Make IO class handle csv file loading for local files mode; IO class hands data to mediator, other classes can get published data from mediator. Move bubbles.prepareData and bubbles.prepareAreas logic to IO class;
