+++ vis/js/bubbles.js
@@ -70,160 +70,6 @@ BubblesFSM.prototype = {
         }
     },
 
-    prepareData: function (init_data, highlight_data) {
-
-        var xy_array = [];
-
-        // convert to numbers
-        init_data.forEach(function (d) {
-            d.x = parseFloat(d.x);
-            d.y = parseFloat(d.y);
-            //if two items have the exact same location, that throws off the force-based layout
-            var xy_string = d.x + d.y;
-            while (xy_array.hasOwnProperty(xy_string)) {
-                d.y += 0.00000001;
-                xy_string = d.x + d.y;
-            }
-
-            xy_array[xy_string] = true;
-            if (d.paper_abstract === null || typeof d.paper_abstract === "undefined") {
-                d.paper_abstract = "";
-            }
-
-            if (d.published_in === null || typeof d.published_in === "undefined") {
-                d.published_in = "";
-            }
-			
-			if (d.title === null || typeof d.title === "undefined") {
-                d.title = headstart.localization[headstart.language]["no_title"];
-            }
-
-            if (headstart.content_based === false) {
-                d.readers = +d.readers;
-                d.internal_readers = +d.readers + 1;
-            } else {
-                d.readers = 0;
-                d.internal_readers = 1;
-            }
-            if (typeof highlight_data != 'undefined' && highlight_data !== null) {
-                if (highlight_data.bookmarks_all !== null) {
-                    highlight_data.bookmarks_all.forEach(function (x) {
-                        var id_string = x.id;
-                        id_string = id_string.toString();
-                        if (id_string == d.id) {
-                            d.readers += x.num;
-                            d.internal_readers += x.num;
-                        }
-                    });
-                }
-            }
-
-            d.paper_selected = false;
-            // convert authors to "[first name] [last name]"
-            var authors = d.authors.split(";");
-            var authors_string = "";
-            var authors_short_string = "";
-            for (var i = 0; i < authors.length - 1; i++) {
-                var names = authors[i].trim().split(",");
-                var last_name = names[0].trim();
-                if (names.length > 1) {
-                    var first_name = names[1].trim();
-                    authors_string += first_name + " " + last_name;
-                    authors_short_string += first_name.substr(0, 1) + ". " + last_name;
-                } else {
-                    authors_string += last_name;
-                    authors_short_string += last_name;
-                }
-
-                if (i != authors.length - 2) {
-                    authors_string += ", ";
-                    authors_short_string += ", ";
-                }
-            }
-
-            d.authors_string = authors_string;
-            d.authors_short_string = authors_short_string;
-            
-            d.oa = false;
-            
-            if (headstart.service === "doaj") {
-                d.oa = true;
-                d.oa_link = d.link;
-            } else if( headstart.service === "plos") {
-                d.oa = true;
-                var journal = d.published_in.toLowerCase();
-                d.oa_link = "http://journals.plos.org/" + headstart.plos_journals_to_shortcodes[journal] + "/article/asset?id=" + d.id + ".PDF";
-            } else if (typeof d.pmcid !== "undefined") {
-                if (d.pmcid !== "") {
-                    d.oa = true;
-                    d.oa_link = "http://www.ncbi.nlm.nih.gov/pmc/articles/" + d.pmcid + "/pdf/";
-                }
-            }
-            
-            d.outlink = list.createOutlink(d);
-
-        });
-        headstart.chart_x.domain(d3.extent(init_data, function (d) {
-            return d.x;
-        }));
-        headstart.chart_y.domain(d3.extent(init_data, function (d) {
-            return d.y * -1;
-        }));
-        headstart.diameter_size.domain(d3.extent(init_data, function (d) {
-            return d.internal_readers;
-        }));
-        var areas = this.areas;
-        init_data.forEach(function (d) {
-
-            // construct rectangles of a golden cut
-            d.diameter = headstart.diameter_size(d.internal_readers);
-            d.width = headstart.paper_width_factor * Math.sqrt(Math.pow(d.diameter, 2) / 2.6);
-            d.height = headstart.paper_height_factor * Math.sqrt(Math.pow(d.diameter, 2) / 2.6);
-            d.orig_x = d.x;
-            d.orig_y = d.y;
-            // scale x and y
-            d.x = headstart.chart_x(d.x);
-            d.y = headstart.chart_y(d.y * -1);
-            var area = (headstart.use_area_uri) ? (d.area_uri) : (d.area);
-            if (area in areas) {
-                areas[area].papers.push(d);
-            } else {
-                areas[area] = {};
-                areas[area].title = d.area;
-                areas[area].papers = [d];
-            }
-
-            d.resized = false;
-        });
-        if (typeof highlight_data != 'undefined' && highlight_data !== null) {
-            if (highlight_data.bookmarks !== null) {
-                highlight_data.bookmarks.forEach(function (d) {
-
-                    var index = init_data.filter(function (x) {
-                        return x.id == d.contentID;
-                    });
-                    if (index.length > 0) {
-                        index[0].bookmarked = 1;
-                    }
-                });
-            }
-
-            if (highlight_data.recommendations !== null) {
-                highlight_data.recommendations.forEach(function (d) {
-
-                    var index = init_data.filter(function (x) {
-                        return x.id == d.contentID;
-                    });
-                    if (index.length > 0) {
-                        index[0].recommended = 1;
-                    }
-                });
-            }
-
-        }
-
-        this.data = init_data;
-    },
     // initialize all "mouseover", "mouseout" event listeners
     // for bubbles
     initMouseListeners: function () {
@@ -463,73 +309,6 @@ BubblesFSM.prototype = {
                 .html(function (d) {
                     return bubbleTemplate(d);
                 });
-    },
-
-    // prepare the areas for the bubbles
-    prepareAreas: function () {
-
-        var areas = this.areas;
-        var areas_array = this.areas_array;
-
-        var readers = [];
-
-        for (var area in areas) {
-            var papers = areas[area].papers;
-            var sum_readers = d3.sum(papers, function (d) {
-                return d.internal_readers;
-            });
-
-            readers.push(sum_readers);
-        }
-
-        headstart.circle_size.domain(d3.extent(readers));
-        var area_x = [];
-        var area_y = [];
-
-        for (area in areas) {
-            let papers = areas[area].papers;
-            let sum_readers = d3.sum(papers, function (d) {
-                return d.internal_readers;
-            });
-
-            areas[area].readers = sum_readers;
-
-            var mean_x = d3.mean(papers, function (d) {
-                return d.x;
-            });
-            var mean_y = d3.mean(papers, function (d) {
-                return d.y;
-            });
-            var r = headstart.circle_size(sum_readers);
-
-            area_x.push(mean_x);
-            area_y.push(mean_y);
-
-            areas[area].x = mean_x;
-            areas[area].y = mean_y;
-            areas[area].r = r;
-        }
-
-        headstart.chart_x_circle.domain(d3.extent(area_x));
-        headstart.chart_y_circle.domain(d3.extent(area_y));
-
-        for (area in areas) {
-            var new_area = [];
-            new_area.title = areas[area].title;
-            new_area.x = headstart.chart_x_circle(areas[area].x);
-            new_area.y = headstart.chart_y_circle(areas[area].y);
-            new_area.orig_x = areas[area].x;
-            new_area.orig_y = areas[area].y;
-            new_area.r = areas[area].r;
-            new_area.height_html = Math.sqrt(Math.pow(areas[area].r, 2) * 2);
-            new_area.width_html = Math.sqrt(Math.pow(areas[area].r, 2) * 2);
-            new_area.x_html = 0 - new_area.width_html / 2;
-            new_area.y_html = 0 - new_area.height_html / 2;
-            new_area.area_uri = area;
-            new_area.readers = areas[area].readers;
-            new_area.papers = areas[area].papers;
-            areas_array.push(new_area);
-        }
     },
 
     zoom: function (d) {
@@ -955,8 +734,6 @@ BubblesFSM.prototype = {
     },
 
     onstart: function (event, from, to, csv, recommendation_data) {
-        this.prepareData(csv, recommendation_data);
-        this.prepareAreas();
 
         if (headstart.is("timeline")) {
             this.draw();
