+++ vis/js/mediator.js
@@ -2,21 +2,28 @@ import Mediator from 'mediator-js';
 import { headstart } from 'headstart';
 import { papers } from 'papers';
 import { list } from 'list';
-import { sortBy } from 'helpers'; 
+import { sortBy } from 'helpers';
+import { io } from 'io';
 
 var MyMediator = function() {
 
     // mediator
+    this.fileData = [];
     this.mediator = new Mediator();
     this.init();
-
-    this.mediator_states = {
-        popup_visible: false
-    };
 };
 
 MyMediator.prototype = {
+    constructor: MyMediator,
     init: function() {
+        // init logic
+        this.mediator.subscribe("headstart_init", this.headstart_init);
+        this.mediator.subscribe("io_init_done", this.io_init_done);
+
+        // data handling
+        this.mediator.subscribe("prepare_data", this.io_prepare_data);
+        this.mediator.subscribe("prepare_areas", this.io_prepare_areas);
+
         // popup
         this.mediator.subscribe("to_timeline", this.to_timeline);
 
@@ -49,6 +56,44 @@ MyMediator.prototype = {
 
     publish: function() {
         this.mediator.publish(...arguments);
+    },
+
+    headstart_init: function (files, input_format) {
+        console.log("HEADSTART INIT");
+        io.init(files, input_format);
+    },
+
+    io_init_done: function (d) {
+        console.log("IO INIT DONE");
+        this.fileData = d;
+    },
+    
+    io_prepare_data: function (highlight_data, cur_fil_num) {
+        console.log("IO PREPARE DATA");
+       io.prepareData(highlight_data, cur_fil_num);
+    },
+
+    io_prepare_areas: function () {
+        console.log("IO PREPARE AREAS");
+        io.prepareAreas();
+    },
+
+    io_prepare_data_done: function (d) {
+        this.preparedData = d;
+    },
+
+    io_get_prepared_data: function (d) {
+        return this.preparedData;
+    },
+
+    io_prepare_areas_done: function (d) {
+        console.log("IO PREPARE AREAS DONE");
+        this.preparedAreas = d;
+        console.log(this.preparedAreas);
+    },
+
+    io_get_prepared_areas: function (d) {
+        return this.preparedAreas;
     },
 
     to_timeline: function() {
